+++ SRMA/Views/Employee/Create.cshtml
@@ -40,7 +40,7 @@
                                         </div>
                                         <h1 class="h5 mb-1">Agrega la información del empleado</h1>
 
-                                        @using (Html.BeginForm("RegisterEmployee", "Usuario", FormMethod.Post, new { @class = "user" }))
+                                        @using (Html.BeginForm("RegisterEmployee", "Employee", FormMethod.Post, new { @class = "user" }))
                                         {
                                             <div class="form-group row">
                                                 <div class="col-sm-6 mb-3 mb-sm-0">
