+++ SRMA/Views/Employee/EmployeeInfo.cshtml
@@ -1,9 +1,99 @@
-﻿
+﻿@model SRMA.Entities.UserEntity
 @{
     ViewData["Title"] = "Información Empleado";
     Layout = "~/Views/Shared/_LayoutAdmin.cshtml";
 }
 
+<html>
+<body>
+
+    <div>
+        <div class="container">
+            <div class="row">
+                <div class="col-12">
+                    <div class="d-flex align-items-center min-vh-100">
+                        <div class="w-100 d-block bg-white shadow-lg rounded my-5">
+                            <div class="row">
+                                <div class="col-lg-5 d-none d-lg-block bg-register rounded-left"></div>
+                                <div class="col-lg-7">
+                                    <div class="p-5">
+                                        <div class="text-center mb-5">
+                                            <a href="index.html" class="text-dark font-size-22 font-family-secondary">
+                                                <i class="mdi mdi-album"></i> <b>SRMA</b>
+                                            </a>
+                                        </div>
+                                        <h1 class="h5 mb-1">Cambiar los datos del empleado</h1>
+                                        <p class="text-muted mb-4">Una vez hayas terminado, utiliza la opción de "Guardar cambios"</p>
+                                        @using (Html.BeginForm("Edit", "Employee", FormMethod.Post, new { @class = "user" }))
+                                        {
+                                            <div class="form-group row">
+
+                                                @Html.HiddenFor(m => m.IdUser)
+
+                                                <div class="col-sm-6 mb-3 mb-sm-0">
+                                                    <label>Nombre</label>
+                                                    @Html.TextBoxFor(m => m.userName, new {@class="form-control form-control-user", @placeholder="Nombre",@type="text"})
+
+                                                </div>
+                                                <div class="col-sm-6">
+                                                    <label>Apellido</label>
+                                                    @Html.TextBoxFor(m => m.lastName, new {@class="form-control form-control-user", @placeholder="Apellido",@type="text"})
+                                                </div>
+
+                                                <div class="form-group">
+                                                    <label>Correo electronico</label>
+                                                    @Html.TextBoxFor(m => m.email, new {@class="form-control form-control-user", @placeholder="Correo electronico",@type="email"})
+                                                </div>
+                                            </div>
+                                            <div class="form-group row">
+                                                <div class="col-sm-6 mb-3 mb-sm-0">
+                                                    <label>Telefono</label>
+                                                    @Html.TextBoxFor(m => m.cellphone, new {@class="form-control form-control-user", @placeholder="Telefono",@type="tel"})
+
+                                                </div>
+                                            <div class="form-group row">
+                                                    <div class="col-sm-6 mb-3 mb-sm-0">
+                                                        <label>Cedula</label>
+                                                        @Html.TextBoxFor(m => m.Id, new {@class="form-control form-control-user", @placeholder="Cedula", @type="int"})
+                                                    </div>
+                                                    
+                                            </div>
+                                           </div>
+                                            <div class="form-group row">
+                                                <div class="col-sm-6 mb-3 mb-sm-0">
+                                                    <a href="@Url.Action("Index", "Employee")" class="btn btn-secondary btn-block waves-effect waves-light">Volver a la página principal</a>
+                                                </div>
+                                                <div class="col-sm-6 mb-3 mb-sm-0">
+                                                    <input type="submit" class="btn btn-success btn-block waves-effect waves-light" value="Guardar cambios" />
+                                                </div>
+                                            </div>
+                                        }
+                                        <div class="form-group">
+                                            @using (Html.BeginForm("DeleteAcc", "Employee", FormMethod.Post))
+                                            {
+                                                @Html.HiddenFor(m => m.IdUser);
+
+                                                <input type="submit" class="btn btn-danger btn-block waves-effect waves-light" value="Eliminar cuenta" />
+                                            }
+
+                                        </div>
+
+                                        <!-- end row -->
+                                    </div> <!-- end .padding-5 -->
+                                </div> <!-- end col -->
+                            </div> <!-- end row -->
+                        </div> <!-- end .w-100 -->
+                    </div> <!-- end .d-flex -->
+                </div> <!-- end col-->
+            </div> <!-- end row -->
+        </div>
+        <!-- end container -->
+    </div>
+    <!-- end page -->
+
+</body>
+</html>
+@*
 <body>
 
     <div>
@@ -85,3 +175,4 @@
         });
     </script>
 </body>
+*@
