+++ SRMA/Views/User/Profile.cshtml
@@ -95,7 +95,10 @@
         <!-- end container -->
     </div>
     <!-- end page -->
-    @*<script>
+    
+</body>
+</html>
+@*<script>
     const miBoton = document.getElementById("miBoton");
 
     miBoton.addEventListener("click", (event) => {
@@ -137,9 +140,6 @@
     });
     });
     </script>*@
-</body>
-</html>
-
